feat: add the ability to post a JSON payload to the Twilio API's
Updating the configuration of a Flex instance requires a POST request to be made with a JSON payload. This change lays the foundations to support this.
The PostJson accepts an interface for the request body (data) instead of just a map as this prevents the computational overhead of having to marshal the struct to a slice of bytes then create a string from the bytes. The JSON string would then be unmarshalled to a map and then marshalled again to a slice of bytes